F.S : New recipe detail page [WIP]
app bar and page header completed
