+++ foodji_ui/lib/l10n/app_en.arb
@@ -6,6 +6,8 @@
   "global_slogan": "Remix your meals to your heart's content",
   "global_filter": "Filter",
 
+  "app_bar_recipe_details": "Recipe details",
+
   "demo_access": "Press anywhere on the screen to access the demo",
   
   "menu_recipes": "Recipes",
