+++ foodji_ui/lib/l10n/app_fr.arb
@@ -6,6 +6,8 @@
   "global_slogan": "Remix ta bouffe comme tu l'aimes",
   "global_filter": "Filtrer",
 
+"app_bar_recipe_details": "Détails de la recette",
+
   "demo_access": "Appuyez sur l'écran pour accéder à la démonstration",
 
   "menu_recipes": "Recettes",
